Tidying up of login pages and navbar copies
