+++ src/components/NavBar/navbar.jsx
@@ -6,36 +6,17 @@ import { Link } from "react-router-dom";
 import { UserContext, NewUserContext, SignUpContext, MobileNavContext } from "../../App";
 import { useContext, useState, useEffect } from "react";
 import Menu from "../../assets/images/logos/menu.svg";
-import { useAuth, logout } from "../../auth";
-import httpClient from "../../httpClient";
+import { useAuth } from "../../auth";
 
 
 
 function NavBar(props) {
   const [logged]=useAuth();
-  // const [isOpened, setIsOpened]  = useContext(UserContext);
+  const [isOpened, setIsOpened]  = useContext(UserContext);
   // const [newUser, setNewUser] = useContext(NewUserContext);
   // const [isSignUp, setIsSignUp] = useContext(SignUpContext);
-  // const [isMobileClicked, setIsMobileClicked] = useContext(MobileNavContext);
+  const [isMobileClicked, setIsMobileClicked] = useContext(MobileNavContext);
 
-  // Function to log out the user
-  function logMeOut() {
-    httpClient({
-      method: "POST",
-      url: "http://localhost:5000/logout"
-    })
-    .then((response) => {
-      logout()
-      alert("You have successfully logged out")
-    }).catch((error) => {
-      if (error.response) {
-        console.log(error.response)
-        console.log(error.response.status)
-        console.log(error.response.error)
-      }
-    })
-
-  }
 
   // Links that logged in user sees
   const LoggedInLinks = () => {
@@ -47,17 +28,16 @@ function NavBar(props) {
             <Link className={styles.link} to="/forums"><h3 className={styles.navItem}>Forums</h3></Link>
             <Link className={styles.link} to="/editAccount"><h3 className={styles.navItem}>Profile</h3></Link>
             <Link className={styles.link} to="/about"><h3 className={styles.navItem}>About</h3></Link>
-            <Link className={styles.link} to="/finalProject"><h3 className={styles.navItem} onClick={logMeOut}>Sign Out</h3></Link>
-            {/* <a className={styles.link} href="/finalProject"><Button text="Sign out" onClick={logMeOut}>Sign out</Button></a> */}
+            {/* <Link className={styles.link} to="/finalProject"><h3 className={styles.navItem} onClick={logMeOut}>Sign Out</h3></Link> */}
             <div className={styles.profileRow}>
             <div className={styles.profile}>
                 <img className={styles.profilePicture} src={profile} alt="profile."/>
             </div>
-            {/* <img 
+            <img 
             className={styles.profilePicture} 
             src={profileDropArrow} 
             alt="drop arrow." 
-            onClick={() => props.onChangePressed(!props.isPressed)}/> */}
+            onClick={() => props.onChangePressed(!props.isPressed)}/>
             </div>
             </div> 
           </>
@@ -65,11 +45,8 @@ function NavBar(props) {
     )
   }
 
-  // Links that logged in user sees
+  // Links that logged out user sees
   const LoggedOutLinks = () => {
-  // can't redeclare the same useContext not sure what to do but I don't think mobile matters so much for our demo!
-  // const [isMobileClicked, setIsMobileClicked] = useContext(MobileNavContext);
-  const [isOpened, setIsOpened]  = useContext(UserContext);
     return (
       <>
         <div className={styles.navItems}>
@@ -84,12 +61,12 @@ function NavBar(props) {
     )
   }
 
-  // Rendering the nav bar
+  // Rendering the nav bar depending on if user is logged in or not
   return(
     <nav className={styles.navbar}>
         <div className={styles.menu}>
         <Link className={styles.link} to="/finalProject"><h1>IntroVerse</h1></Link>
-        {/* <img className={styles.MobileMenu} src={Menu} alt="dropDownMenu" onClick={() => setIsMobileClicked(!isMobileClicked)}/> */}
+        <img className={styles.MobileMenu} src={Menu} alt="dropDownMenu" onClick={() => setIsMobileClicked(!isMobileClicked)}/>
         </div>
         {logged ? <LoggedInLinks /> : <LoggedOutLinks />}
     </nav>
